config/requirements.txt
# Configuration system requirements
PyYAML>=6.0
